Factor shared script logic out into `utils` module
